requirements.txt
streamlit==1.34.0
pandas==2.1.4
numpy==1.26.0
pickle-mixin==1.0.2
scikit-learn==1.4.2
tensorflow==2.17.1
gspread
gspread-dataframe
google-auth
google-auth-oauthlib
oauth2client
